pyproject.toml
[tool.poetry]
name = "zonehelper"
version = "0.1.0"
description = ""
authors = ["Aleksandr Efimovskiy <bdf715@gmail.com>"]

packages = [
  { include = "zonehelper" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
prompt = "^0.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
zonehelper = "zonehelper.scripts.zonehelper:main"
